Move ClusterGetter.get to Cluster.cluster_ocns!
* remove HtItemClusterGetter -- we
* remove HtItemCluster functionality (for now) since we aren't
  persisting anything about HtItems except that their OCNs should be in
  a cluster
* move ClusterGetter specs to cluster; re-do the
  instrumentation/concurrency specs using thread-local storage
  for storing "cluster instrumentation"
