+++ lib/clustering/cluster_ht_item.rb
@@ -2,8 +2,6 @@
 
 require "cluster"
 require "clustering/reclusterer"
-require "clustering/cluster_getter"
-require "clustering/ht_item_cluster_getter"
 
 module Clustering
   # Services for batch loading HT items
@@ -22,79 +20,16 @@ module Clustering
       end
     end
 
-    def cluster(getter: cluster_getter)
-      # For the case where there are no OCNs, ClusterGetter always creates a new
-      # cluster. We might want to first consider searching by item ID when there
-      # are no OCNs.
+    def cluster
+      return if ocns.empty?
 
-      getter.get do |cluster|
-        update_or_add_ht_items(cluster)
-      end
-    end
-
-    # Removes an HTItem
-    def delete
-      raise "not implemented"
-      raise ArgumentError, "Can only delete one HTItem at a time" unless ht_items.length == 1
-
-      ht_item = ht_items.first
-      # Don't start a transaction until we know there's something we need to
-      # delete. If we do need to delete the thing, then we need to re-fetch it so
-      # we can acquire the correct lock
-      return unless cluster_with_htitem(ht_item)
-
-      Retryable.with_transaction do
-        if (cluster = cluster_with_htitem(ht_item))
-          Services.logger.debug "removing old htitem #{ht_item.item_id}"
-          old_item = cluster.ht_item(ht_item.item_id)
-          old_item.delete
-
-          Reclusterer.new(cluster, old_item.ocns).recluster
-        end
-      end
+      # Right now, we aren't persisting anything except ensuring that the OCNs
+      # for the htitem are in the same cluster
+      Cluster.cluster_ocns!(ocns)
     end
 
     private
 
     attr_reader :ht_items, :ocns
-
-    def cluster_getter
-      if ocns.empty?
-        Clustering::HtItemClusterGetter.new(*ht_items)
-      else
-        ClusterGetter.new(ocns)
-      end
-    end
-
-    def cluster_with_htitem(htitem)
-      Cluster.with_ht_item(htitem).first
-    end
-
-    def update_or_add_ht_items(cluster)
-      Services.logger.debug "Cluster #{cluster.inspect}: " \
-        "adding ht_items #{ht_items.inspect} with ocns #{ocns}"
-      to_append = []
-      ht_items.each do |item|
-        if (existing_item = cluster.ht_item(item.item_id))
-          Services.logger.debug "updating existing item with id #{item.item_id}"
-
-          if (item_attrs = item.to_hash) != existing_item.to_hash
-            existing_item.update_attributes(item_attrs)
-            @any_updated = true
-          end
-        else
-          ClusterHtItem.new(item).delete
-          to_append << item
-        end
-      end
-
-      cluster.add_ht_items(to_append) unless to_append.empty?
-      if @any_updated || to_append.any?
-        # Reclusterer would handle this if we could tell it which tuples were removed.
-        cluster.update_ocns
-        cluster.save
-        Reclusterer.new(cluster).recluster
-      end
-    end
   end
 end
